Fixed Tailwindcss cdn option +  Separated theme script for ease of toggling in ShadHead()
