+++ shad4fast/src/shad4fast/__about__.py
@@ -1,4 +1,4 @@
 # SPDX-FileCopyrightText: 2024-present Curtis Allan <curtis@Curtiss-MacBook-Air.local>
 #
 # SPDX-License-Identifier: MIT
-__version__ = "0.1"
+__version__ = "0.2"
